// types.ts
export enum CaseType {
  CIVIL_GENERAL = 'CIVIL_GENERAL',
  BANKING_CONSUMER = 'BANKING_CONSUMER',
  PAULIAN_ACTION = 'PAULIAN_ACTION',
  LABOR_EMPLOYEE = 'LABOR_EMPLOYEE',
  ALIMONY = 'ALIMONY',
  DIVORCE = 'DIVORCE',
  SEPARATION = 'SEPARATION',
  PROTECTION_PERSONAL_RIGHTS = 'PROTECTION_PERSONAL_RIGHTS',
  MARRIAGE_PERMISSION = 'MARRIAGE_PERMISSION',
  MATRIMONIAL_PROPERTY_DISSOLUTION = 'MATRIMONIAL_PROPERTY_DISSOLUTION',
  ESTATE_DIVISION_JOINT = 'ESTATE_DIVISION_JOINT',
  ESTATE_DIVISION_JOINT_AGREED = 'ESTATE_DIVISION_JOINT_AGREED',
  REAL_ESTATE_ZASIEDZENIE = 'REAL_ESTATE_ZASIEDZENIE',
  CO_OWNERSHIP_DISSOLUTION = 'CO_OWNERSHIP_DISSOLUTION',
  CO_OWNERSHIP_DISSOLUTION_AGREED = 'CO_OWNERSHIP_DISSOLUTION_AGREED',
  EVICTION_RESIDENTIAL = 'EVICTION_RESIDENTIAL',
  POSSESSION_DISTURBANCE = 'POSSESSION_DISTURBANCE',
  LAND_REGISTER_OWNERSHIP = 'LAND_REGISTER_OWNERSHIP',
  LAND_REGISTER_MORTGAGE = 'LAND_REGISTER_MORTGAGE',
  INHERITANCE_STATEMENT = 'INHERITANCE_STATEMENT',
  INHERITANCE_DIVISION = 'INHERITANCE_DIVISION',
  INHERITANCE_DIVISION_AGREED = 'INHERITANCE_DIVISION_AGREED',
  KRS_REGISTRATION = 'KRS_REGISTRATION',
  KRS_CHANGE = 'KRS_CHANGE',
  COMPANY_DISSOLUTION = 'COMPANY_DISSOLUTION',
  PRIVATIZATION_TENDER = 'PRIVATIZATION_TENDER',
}

export enum ProcedureType {
  STANDARD = 'STANDARD',
  WRIT_PROCEEDINGS = 'WRIT_PROCEEDINGS',
  ORDER_PAYMENT_ELECTRONIC = 'ORDER_PAYMENT_ELECTRONIC',
  APPEAL = 'APPEAL',
  PRIVATE_MEDIATION_SETTLEMENT = 'PRIVATE_MEDIATION_SETTLEMENT'
}

export interface CalculationResult {
  fee: number;
  description: string;
  legalBasis: string;
}